update code and pickers added
