+++ lua/Arduino-Nvim/init.lua
@@ -1,8 +1,3 @@
-local remap = require("Arduino-Nvim.remap")
-local commands = require("Arduino-Nvim.commands")
-local b_config = require("Arduino-Nvim.board_config")
-local gui = require("Arduino-Nvim.gui")
-local lsp = require("Arduino-Nvim.lsp")
 local M = {}
 
 -- Default config values
@@ -11,22 +6,23 @@ _ArduinoConfigValues = {
   board = "arduino:avr:uno",
   port = "/dev/ttyUSB0",
   baudrate = 115200,
+  picker_backend = "telescope",
   use_default_keymaps = true,
   use_default_commands = true,
   keymaps = {},
 }
 
-local function set_default_commands()
+local function set_default_commands(commands, gui, b_config, picker)
   vim.api.nvim_create_user_command("InoMonitor",     commands.monitor, {})
   vim.api.nvim_create_user_command("InoUpload",      commands.upload, {})
   vim.api.nvim_create_user_command("InoUploadSlow",  commands.upload_slow, {})
   vim.api.nvim_create_user_command("InoUploadReset", commands.upload_reset, {})
   vim.api.nvim_create_user_command("InoDebugUpload", commands.upload_debug, {})
-  -- vim.api.nvim_create_user_command("InoGUI",         gui.gui, {})
-  -- vim.api.nvim_create_user_command("InoList",        gui.arduino_board_list_gui, {})
-  -- vim.api.nvim_create_user_command("InoLib",         gui.library_manager_gui, {})
-  -- vim.api.nvim_create_user_command("InoSelectBoard", gui.select_board_gui, {})
-  -- vim.api.nvim_create_user_command("InoSelectPort",  gui.select_port_gui, {})
+  vim.api.nvim_create_user_command("InoList",        gui.arduino_board_list_gui, {})
+  vim.api.nvim_create_user_command("InoGUI",         picker.select_board_and_port, {})
+  vim.api.nvim_create_user_command("InoLib",         picker.open_library_manager, {})
+  vim.api.nvim_create_user_command("InoSelectBoard", picker.select_board, {})
+  vim.api.nvim_create_user_command("InoSelectPort",  picker.select_port, {})
   vim.api.nvim_create_user_command("InoStatus",      b_config.board_config_status, {})
 
   vim.api.nvim_create_user_command("InoCheck",       function()
@@ -52,6 +48,12 @@ local function update_config(opts)
 end
 
 function M.setup(opts)
+  local b_config = require("Arduino-Nvim.board_config")
+  local commands = require("Arduino-Nvim.commands")
+  local gui = require("Arduino-Nvim.gui")
+  local lsp = require("Arduino-Nvim.lsp")
+  local remap = require("Arduino-Nvim.remap")
+
   -- Update global config values
   update_config(opts)
 
@@ -66,7 +68,12 @@ function M.setup(opts)
 
   -- load default commands
   if _ArduinoConfigValues.use_default_commands then
-    set_default_commands()
+    set_default_commands(
+      commands,
+      gui,
+      b_config,
+      require("Arduino-Nvim.pickers.".._ArduinoConfigValues.picker_backend)
+    )
   end
 end
